db: Introduce dbtest and extend dbutil packages
